requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
albumentations>=1.3.0
matplotlib>=3.7.0
pandas>=2.0.0
tqdm>=4.65.0
jupyter>=1.0.0
pyyaml>=6.0.0
